Sort events based on date
For whatever reason, Jekyll doesn't seem to like piping after sorting. Unfortunately,
we want the sort order to be reverse chronological. In order to get around this I
first assign the events variable to the chronological ordering with sort: 'date',
and then specify reversed when iterating over events within the for loop.