Fix erroneous Sass @extend-inside-@media expressions
In general, @extend expressions are not allowed inside @media blocks.

This will not compile with the latest version of Sass. It works with
node-sass because node-sass uses libsass, but libsass has a bug that
allows this even though it is not syntactically valid.

This was a blocker from jettisoning the Node toolchain and going to
Rake.